DAOS-8040 control: enable vmd support by caching state in bdev forwarder
VMD state is detected during NVMe storage prepare phase, cache this
state in the BdevAdminForwarder which handles the forwarding of requests
from daos_server to daos_admin process instances. The state can then be
used to determine whether or not to enable VMD processing in bdev backend.
