+++ src/control/server/ctl_storage.go
@@ -8,14 +8,12 @@ package server
 
 import (
 	"context"
-	"fmt"
 	"path/filepath"
 	"strings"
 
 	"github.com/dustin/go-humanize"
 	"github.com/pkg/errors"
 
-	"github.com/daos-stack/daos/src/control/common"
 	"github.com/daos-stack/daos/src/control/logging"
 	"github.com/daos-stack/daos/src/control/server/engine"
 	"github.com/daos-stack/daos/src/control/server/storage"
@@ -60,127 +58,6 @@ func NewMockStorageControlService(log logging.Logger, engineCfgs []*engine.Confi
 	)
 }
 
-// findBdevsWithDomain retrieves controllers in scan response that match the
-// input prefix in the domain component of their PCI address.
-func findBdevsWithDomain(scanResp *storage.BdevScanResponse, prefix string) ([]string, error) {
-	var found []string
-
-	for _, ctrlr := range scanResp.Controllers {
-		domain, _, _, _, err := common.ParsePCIAddress(ctrlr.PciAddr)
-		if err != nil {
-			return nil, err
-		}
-		if fmt.Sprintf("%x", domain) == prefix {
-			found = append(found, ctrlr.PciAddr)
-		}
-	}
-
-	return found, nil
-}
-
-// substBdevVmdAddrs replaces VMD PCI addresses in bdev device list with the
-// PCI addresses of the backing devices behind the VMD.
-//
-// Select any PCI addresses that have the compressed VMD address BDF as backing
-// address domain.
-//
-// Return new device list with PCI addresses of devices behind the VMD.
-func substBdevVmdAddrs(cfgBdevs []string, scanResp *storage.BdevScanResponse) ([]string, error) {
-	if len(scanResp.Controllers) == 0 {
-		return nil, nil
-	}
-
-	var newCfgBdevs []string
-	for _, dev := range cfgBdevs {
-		_, b, d, f, err := common.ParsePCIAddress(dev)
-		if err != nil {
-			return nil, err
-		}
-		matchDevs, err := findBdevsWithDomain(scanResp,
-			fmt.Sprintf("%02x%02x%02x", b, d, f))
-		if err != nil {
-			return nil, err
-		}
-		if len(matchDevs) == 0 {
-			newCfgBdevs = append(newCfgBdevs, dev)
-			continue
-		}
-		newCfgBdevs = append(newCfgBdevs, matchDevs...)
-	}
-
-	return newCfgBdevs, nil
-}
-
-// canAccessBdevs evaluates if any specified Bdevs are not accessible.
-//
-// Specified Bdevs can be VMD addresses.
-//
-// Return any device addresses missing from the scan response and ok set to true
-// if no devices are missing.
-func canAccessBdevs(cfgBdevs []string, scanResp *storage.BdevScanResponse) ([]string, bool) {
-	var missing []string
-
-	for _, pciAddr := range cfgBdevs {
-		var found bool
-
-		for _, ctrlr := range scanResp.Controllers {
-			if ctrlr.PciAddr == pciAddr {
-				found = true
-				break
-			}
-		}
-		if !found {
-			missing = append(missing, pciAddr)
-		}
-	}
-
-	return missing, len(missing) == 0
-}
-
-// checkCfgBdevs performs validation on NVMe returned from initial scan.
-func (c *StorageControlService) checkCfgBdevs(scanResp *storage.BdevScanResponse) error {
-	if scanResp == nil || scanResp.Controllers == nil {
-		return errors.New("received nil scan response or controllers")
-	}
-	if len(c.instanceStorage) == 0 {
-		return nil
-	}
-
-	for _, storageCfg := range c.instanceStorage {
-		for _, tierCfg := range storageCfg.Tiers {
-			if !tierCfg.IsBdev() || len(tierCfg.Bdev.DeviceList) == 0 {
-				continue
-			}
-			cfgBdevs := tierCfg.Bdev.DeviceList
-
-			// TODO DAOS-8040: re-enable VMD
-			// if !c.bdev.IsVMDDisabled() {
-			// 	c.log.Debug("VMD detected, processing PCI addresses")
-			// 	newBdevs, err := substBdevVmdAddrs(cfgBdevs, scanResp)
-			// 	if err != nil {
-			// 		return err
-			// 	}
-			// 	if len(newBdevs) == 0 {
-			// 		return errors.New("unexpected empty bdev list returned " +
-			// 			"check vmd address has backing devices")
-			// 	}
-			// 	c.log.Debugf("instance %d: subst vmd addrs %v->%v",
-			// 		idx, cfgBdevs, newBdevs)
-			// 	cfgBdevs = newBdevs
-			// 	c.instanceStorage[idx].Bdev.DeviceList = cfgBdevs
-			// }
-
-			// fail if config specified nvme devices are inaccessible
-			missing, ok := canAccessBdevs(cfgBdevs, scanResp)
-			if !ok {
-				return FaultBdevNotFound(missing)
-			}
-		}
-	}
-
-	return nil
-}
-
 // Setup delegates to Storage implementation's Setup methods.
 func (c *StorageControlService) Setup() error {
 	if _, err := c.ScmScan(storage.ScmScanRequest{}); err != nil {
@@ -197,14 +74,12 @@ func (c *StorageControlService) Setup() error {
 		}
 	}
 
-	nvmeScanResp, err := c.NvmeScan(storage.BdevScanRequest{})
+	nvmeScanResp, err := c.NvmeScan(storage.BdevScanRequest{
+		EngineStorage: c.instanceStorage,
+	})
 	if err != nil {
 		c.log.Debugf("%s\n", errors.Wrap(err, "Warning, NVMe Scan"))
 		return nil
-	}
-
-	if err := c.checkCfgBdevs(nvmeScanResp); err != nil {
-		return errors.Wrap(err, "validate server config bdevs")
 	}
 
 	c.storage.SetBdevCache(*nvmeScanResp)
