+++ src/control/server/faults.go
@@ -13,7 +13,6 @@ import (
 	"github.com/dustin/go-humanize"
 
 	"github.com/daos-stack/daos/src/control/build"
-	"github.com/daos-stack/daos/src/control/common"
 	"github.com/daos-stack/daos/src/control/fault"
 	"github.com/daos-stack/daos/src/control/fault/code"
 	"github.com/daos-stack/daos/src/control/server/engine"
@@ -126,14 +125,6 @@ func FaultScmUnmanaged(mntPoint string) *fault.Fault {
 		code.ServerScmUnmanaged,
 		fmt.Sprintf("the SCM mountpoint at %s is unavailable and can't be created/mounted", mntPoint),
 		fmt.Sprintf("manually create %s or remove --recreate-superblocks from the server arguments", mntPoint),
-	)
-}
-
-func FaultBdevNotFound(bdevs []string) *fault.Fault {
-	return serverFault(
-		code.ServerBdevNotFound,
-		fmt.Sprintf("NVMe SSD%s %v not found", common.Pluralise("", len(bdevs)), bdevs),
-		fmt.Sprintf("check SSD%s %v that are specified in server config exist", common.Pluralise("", len(bdevs)), bdevs),
 	)
 }
