+++ src/control/server/storage/bdev/provider.go
@@ -22,9 +22,6 @@ type (
 		Reset(storage.BdevPrepareRequest) (*storage.BdevPrepareResponse, error)
 		Scan(storage.BdevScanRequest) (*storage.BdevScanResponse, error)
 		Format(storage.BdevFormatRequest) (*storage.BdevFormatResponse, error)
-		// TODO DAOS-8040: re-enable VMD
-		//		EnableVMD()
-		//		IsVMDEnabled() bool
 		UpdateFirmware(pciAddr string, path string, slot int32) error
 		WriteNvmeConfig(storage.BdevWriteNvmeConfigRequest) (*storage.BdevWriteNvmeConfigResponse, error)
 	}
@@ -51,25 +48,9 @@ func NewProvider(log logging.Logger, backend Backend) *Provider {
 	return p
 }
 
-// TODO DAOS-8040: re-enable VMD
-//func (p *Provider) enableVMD() {
-//	p.backend.EnableVMD()
-//}
-
-// IsVMDEnabled returns true if provider has enabled VMD device awareness.
-//func (p *Provider) IsVMDEnabled() bool {
-//	return p.backend.IsVMDEnabled()
-//}
-
 // Scan calls into the backend to discover NVMe components in the
 // system.
 func (p *Provider) Scan(req storage.BdevScanRequest) (resp *storage.BdevScanResponse, err error) {
-	// TODO DAOS-8040: re-enable VMD
-	// set vmd state on remote provider in forwarded request
-	//	if req.IsForwarded() && req.EnableVMD {
-	//		p.enableVMD()
-	//	}
-
 	return p.backend.Scan(req)
 }
 
@@ -93,12 +74,6 @@ func (p *Provider) Format(req storage.BdevFormatRequest) (*storage.BdevFormatRes
 	if len(req.Properties.DeviceList) == 0 {
 		return nil, errors.New("empty DeviceList in FormatRequest")
 	}
-
-	// TODO DAOS-8040: re-enable VMD
-	// set vmd state on remote provider in forwarded request
-	//	if req.IsForwarded() && req.EnableVMD {
-	//		p.enableVMD()
-	//	}
 
 	return p.backend.Format(req)
 }
